Move internal state to presentational component
