+++ client/src/components/v2/OverlayMessageContainer.tsx
@@ -1,5 +1,5 @@
 import { sendMessage } from 'api/v2/camera';
-import React, { useCallback, useState } from 'react';
+import React from 'react';
 import OverlayMessage from './OverlayMessage';
 
 /**
@@ -8,28 +8,7 @@ import OverlayMessage from './OverlayMessage';
  * @returns Component
  */
 export default function OverlayMessageContainer() {
-    const [message, setMessage] = useState('');
-
-    const handleMessageChange = useCallback(
-        (event) => setMessage(event.target.value),
-        [setMessage],
-    );
-
-    const handleKeyPressed = useCallback(
-        (event) => {
-            if (event.key === 'Enter') {
-                sendMessage(message);
-                setMessage('');
-            }
-        },
-        [message, setMessage],
-    );
-
     return (
-        <OverlayMessage
-            message={message}
-            handleMessageChange={handleMessageChange}
-            handleKeyPressed={handleKeyPressed}
-        />
+        <OverlayMessage sendMessage={sendMessage} />
     );
 }
